chore: more types and wrap controls
