+++ front/src/features/home/SessionController.js
@@ -1,34 +1,41 @@
-import React, { Component } from 'react';
-import PropTypes from 'prop-types';
-import { bindActionCreators } from 'redux';
+import React from 'react';
 import { connect } from 'react-redux';
+import { makeStyles } from '@material-ui/core/styles';
 import * as actions from './redux/actions';
 import SessionStart from './SessionStart';
 import SessionStop from './SessionStop';
 
-export class SessionController extends Component {
-  static propTypes = {
-    home: PropTypes.object.isRequired,
-    actions: PropTypes.object.isRequired,
-  };
+const useStyles = makeStyles(theme => ({
+  root: {
+    display: 'flex',
+    justifyContent: 'center',
+    flexDirection: 'column',
+    alignItems: 'center',
+  },
+}));
 
-  handleChange = event => this.props.actions.setSessionName(event.target.value);
-
-  render() {
-    const { runningSession } = this.props.home;
-    const { startSession, stopSession } = this.props.actions;
-    return runningSession ? (
-      <SessionStop
-        stopSession={stopSession}
-        name={runningSession.name}
-        id={runningSession.id}
-        startDate={runningSession.startDate}
-      />
-    ) : (
-      <SessionStart startSession={startSession} />
-    );
-  }
-}
+/**
+ * @param {import('react-redux').ConnectedProps<typeof connector>} props
+ */
+export const SessionController = props => {
+  const css = useStyles();
+  const { runningSession } = props.home;
+  const { startSession, stopSession } = props;
+  return (
+    <div className={css.root}>
+      {runningSession ? (
+        <SessionStop
+          stopSession={stopSession}
+          name={runningSession.name}
+          id={runningSession.id}
+          startDate={runningSession.startDate}
+        />
+      ) : (
+        <SessionStart startSession={startSession} />
+      )}
+    </div>
+  );
+};
 
 /* istanbul ignore next */
 /**
@@ -40,11 +47,6 @@ function mapStateToProps(state) {
   };
 }
 
-/* istanbul ignore next */
-function mapDispatchToProps(dispatch) {
-  return {
-    actions: bindActionCreators({ ...actions }, dispatch),
-  };
-}
+const connector = connect(mapStateToProps, actions);
 
-export default connect(mapStateToProps, mapDispatchToProps)(SessionController);
+export default connector(SessionController);
