+++ front/src/features/home/redux/editSession.js
@@ -2,18 +2,23 @@
 // putting related actions and reducers in one file. See more at:
 // https://medium.com/@nate_wang/a-new-approach-for-managing-redux-actions-91c26ce8b5da
 
-import {
-  HOME_EDIT_SESSION,
-} from './constants';
+import { HOME_EDIT_SESSION } from './constants';
 
+/**
+ * @param {string} id
+ */
 export function editSession(id) {
   return {
     type: HOME_EDIT_SESSION,
-    payload: { id }
+    payload: { id },
   };
 }
 
-export function reducer(state, action = { payload: {} }) {
+/**
+ * @param {import('./types').InitialState} state
+ * @param {{type: HOME_EDIT_SESSION, payload: {id: string}}} action
+ */
+export function reducer(state, action) {
   switch (action.type) {
     case HOME_EDIT_SESSION:
       return {
